Abstract plugin, basic model plugin
