+++ klum-ast-gradle-plugin/src/main/java/com/blackbuild/klum/ast/gradle/KlumAstModelPlugin.java
@@ -1,0 +1,16 @@
+package com.blackbuild.klum.ast.gradle;
+
+import org.gradle.api.NonNullApi;
+
+@NonNullApi
+public class KlumAstModelPlugin extends AbstractKlumPlugin {
+
+    @Override
+    protected void addDependencies() {
+
+    }
+
+    protected void addDependentPlugins() {
+        // no additional dependencies
+    }
+}
